requirements.txt
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
pandas==2.0.3
numpy==1.24.3
schedule==1.2.0
flask-cors==4.0.0
